feat(kernel): UAC2 async audio OUT with debug instrumentation
- Add UAC2 mode selection via bsp.hh defines
- Extensive debug counters for USB audio diagnostics
- PDM mic integration with CIC decimation
- Update synth_app and build system for new kernel interface

+++ examples/stm32f4_kernel/src/kernel.hh
@@ -13,7 +13,7 @@ void on_audio_buffer_ready(uint16_t* buf);
 void on_pdm_buffer_ready(uint16_t* buf);
 
 // Tick counter (microseconds)
-extern volatile uint32_t g_tick_us;
+extern volatile uint64_t g_tick_us;
 
 // Current sample rate
 extern volatile uint32_t g_current_sample_rate;
